requirements.txt
torch
web3
py-solc-x
numpy
streamlit
